+++ config/Migrations/20190208174112_AddAdditionalDataToUsers.php
@@ -1,0 +1,32 @@
+<?php
+/**
+ * Copyright 2010 - 2019, Cake Development Corporation (https://www.cakedc.com)
+ *
+ * Licensed under The MIT License
+ * Redistributions of files must retain the above copyright notice.
+ *
+ * @copyright Copyright 2010 - 2019, Cake Development Corporation (https://www.cakedc.com)
+ * @license MIT License (http://www.opensource.org/licenses/mit-license.php)
+ */
+
+use Migrations\AbstractMigration;
+
+class AddAdditionalDataToUsers extends AbstractMigration
+{
+    /**
+     * Change Method.
+     *
+     * More information on this method is available here:
+     * http://docs.phinx.org/en/latest/migrations.html#the-change-method
+     * @return void
+     */
+    public function change()
+    {
+        $table = $this->table('users');
+        $table->addColumn('additional_data', 'text', [
+            'default' => null,
+            'null' => true,
+        ]);
+        $table->update();
+    }
+}