+++ CHANGELOG.md
@@ -4,6 +4,26 @@ Changelog
 Releases for CakePHP 3
 -------------
 
+* 8.4.0
+  * Rehash password if needed at login
+  
+* 8.3.0
+  * Bootstrap don't need to listen for EVENT_FAILED_SOCIAL_LOGIN
+
+* 8.2.1
+  * Fix scope in facebook social login
+
+* 8.2
+  * Removed deprecations for CakePHP 3.7
+
+* 8.1
+  * Added Yubico U2F Authentication
+
+* 8.0.3
+  * Updated to latest version of Google OAuth
+  * Added plugin object
+  * Fixed action changePassword to work with post or put request
+ 
 * 8.0.2
   * Add default role for users registered via social login